Move extra protocols out of StoreProtocol's file.
